requirements.txt

fastapi==0.110.0
uvicorn==0.30.0
pandas==2.3.2
scikit-learn==1.7.2
joblib==1.4.2
pydantic==2.8.2
sqlalchemy>=2.0.36
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.0
xgboost==2.0.3
numpy==1.26.4
